Moved test files to sub directory
